# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FoodHelper API
  version: v1
  description: API for managing products, recipes, and nutritional calculations.
servers:
  - url: /
    description: Current server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/recipes:
    get:
      tags:
        - Recipes
      summary: List recipes
      description: Returns a paginated list of recipes ordered by identifier ascending, including calculated nutritional totals, ingredients, and optional derived product information.
      operationId: listRecipes
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-based page number
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          description: Number of items per page, between 1 and 100
          schema:
            type: integer
            format: int32
            default: 20
        - name: search
          in: query
          required: false
          description: Accent- and case-insensitive text searched in name, description, instructions, and ingredients
          schema:
            type: string
        - name: caloriesMin
          in: query
          required: false
          schema: { type: number, minimum: 0 }
        - name: caloriesMax
          in: query
          required: false
          schema: { type: number, minimum: 0 }
        - name: carbohydratesMin
          in: query
          required: false
          schema: { type: number, minimum: 0 }
        - name: carbohydratesMax
          in: query
          required: false
          schema: { type: number, minimum: 0 }
        - name: proteinsMin
          in: query
          required: false
          schema: { type: number, minimum: 0 }
        - name: proteinsMax
          in: query
          required: false
          schema: { type: number, minimum: 0 }
        - name: fatsMin
          in: query
          required: false
          schema: { type: number, minimum: 0 }
        - name: fatsMax
          in: query
          required: false
          schema: { type: number, minimum: 0 }
        - name: hasDerivedProduct
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Recipes returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePageResponse'
        '400':
          description: Invalid pagination, negative range, or minimum greater than maximum
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
        - Recipes
      summary: Create recipe
      description: Creates a recipe from existing products and calculates nutritional totals from assigned grams.
      operationId: createRecipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
            examples:
              default:
                value:
                  name: Curry
                  description: Creamy homemade curry
                  instructions: Cook all ingredients together.
                  products:
                    - productId: 1
                      grams: 200
                    - productId: 2
                      grams: 100
      responses:
        '201':
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: One or more ingredient products were not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflict while creating the recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/recipes/{id}:
    put:
      tags:
        - Recipes
      summary: Update recipe
      description: Updates a recipe, recalculates nutritional totals, and synchronizes its derived product if it exists.
      operationId: updateRecipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecipeRequest'
            examples:
              default:
                value:
                  name: Curry
                  description: Creamy homemade curry
                  instructions: Cook all ingredients together.
                  products:
                    - productId: 1
                      grams: 250
                    - productId: 2
                      grams: 100
      responses:
        '200':
          description: Recipe updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Recipe or one of its ingredient products was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflict while updating the recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Recipes
      summary: Delete recipe
      description: Deletes a recipe. If a derived product exists, it is deleted too.
      operationId: deleteRecipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Recipe deleted
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/recipes/{id}/derived-product:
    post:
      tags:
        - Recipes
      summary: Create derived product from recipe
      description: Creates a product from a recipe once and stores the yield metadata used to calculate produced units.
      operationId: createDerivedProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeDerivedProductRequest'
            examples:
              default:
                value:
                  producedGrams: 400
                  gramsPerUnit: 100
      responses:
        '201':
          description: Derived product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDerivedProductResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Derived product already exists or product name conflicts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/users/{personId}/menu-history/monthly:
    get:
      tags: [Users]
      summary: Get monthly menu history
      description: Returns immutable closed-menu snapshots assigned to one person during a month.
      operationId: getMonthlyUserMenuHistory
      parameters:
        - {name: personId, in: path, required: true, schema: {type: integer, format: int64}}
        - {name: year, in: query, required: true, schema: {type: integer, example: 2026}}
        - {name: month, in: query, required: true, schema: {type: integer, minimum: 1, maximum: 12}}
      responses:
        '200':
          description: Monthly history returned
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserMenuHistoryResponse'}
        '404':
          description: Person not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
  /api/v1/users/{personId}/menu-history/annual:
    get:
      tags: [Users]
      summary: Get annual menu history
      description: Returns immutable closed-menu snapshots assigned to one person during a year.
      operationId: getAnnualUserMenuHistory
      parameters:
        - {name: personId, in: path, required: true, schema: {type: integer, format: int64}}
        - {name: year, in: query, required: true, schema: {type: integer, example: 2026}}
      responses:
        '200':
          description: Annual history returned
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserMenuHistoryResponse'}
        '404':
          description: Person not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid request body
        path:
          type: string
          example: /api/v1/products
        timestamp:
          type: string
          format: date-time
...
    CreateRecipeRequest:
      type: object
      required:
        - name
        - description
        - instructions
        - products
      properties:
        name:
          type: string
          minLength: 1
          description: Recipe name
          example: Curry
        description:
          type: string
          minLength: 1
          description: Recipe description
          example: Creamy homemade curry
        instructions:
          type: string
          minLength: 1
          description: Recipe preparation instructions
          example: Cook all ingredients together.
        products:
          type: array
          minItems: 1
          description: Assigned ingredient products with their gram amount
          items:
            $ref: '#/components/schemas/RecipeIngredientAssignmentRequest'
        photo:
          $ref: '#/components/schemas/PhotoUploadRequest'
    UpdateRecipeRequest:
      allOf:
        - $ref: '#/components/schemas/CreateRecipeRequest'
    RecipeIngredientAssignmentRequest:
      type: object
      required:
        - productId
        - grams
      properties:
        productId:
          type: integer
          format: int64
          description: Ingredient product identifier
          example: 1
        grams:
          type: number
          exclusiveMinimum: 0
          description: Assigned grams from the ingredient product
          example: 125
    CreateRecipeDerivedProductRequest:
      type: object
      required:
        - producedGrams
        - gramsPerUnit
      properties:
        producedGrams:
          type: number
          exclusiveMinimum: 0
          description: Total grams of product yielded by the recipe
          example: 400
        gramsPerUnit:
          type: number
          exclusiveMinimum: 0
          description: Grams represented by one product unit
          example: 100
    RecipeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Curry
        description:
          type: string
          example: Creamy homemade curry
        instructions:
          type: string
          example: Cook all ingredients together.
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
        products:
          type: array
          items:
            $ref: '#/components/schemas/RecipeIngredientResponse'
        derivedProduct:
          oneOf:
            - $ref: '#/components/schemas/RecipeDerivedProductResponse'
            - type: 'null'
        photo:
          oneOf:
            - type: string
              description: Signed photo URL that expires with the authentication token lifetime
              example: /api/v1/media/12?expiresAt=1781611200&signature=5f2a...
            - type: 'null'
    RecipePageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecipeResponse'
        page:
          type: integer
          format: int32
          description: Zero-based page number
          example: 0
        size:
          type: integer
          format: int32
          description: Requested page size
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Total number of recipes
          example: 42
        totalPages:
          type: integer
          format: int32
          description: Total number of pages
          example: 3
...
    PhotoUploadRequest:
      type: object
      required:
        - fileName
        - contentType
        - base64Data
      properties:
        fileName:
          type: string
          description: Original file name
          example: apple.png
        contentType:
          type: string
          description: Original MIME type reported by the client
          example: image/png
        base64Data:
          type: string
          description: Base64-encoded image binary
          example: iVBORw0KGgoAAAANSUhEUgAA...
    RecipeIngredientResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 1
        productName:
          type: string
          example: Chicken breast
        grams:
          type: number
          example: 125
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
    RecipeDerivedProductResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 7
        producedGrams:
          type: number
          example: 400
        gramsPerUnit:
          type: number
          example: 100
        unitsProduced:
          type: number
          example: 4
...
    UserMenuHistoryResponse:
      type: object
      properties:
        personId: {type: integer, format: int64, example: 1}
        personName: {type: string, example: elias}
        year: {type: integer, example: 2026}
        month: {type: integer, nullable: true, example: 6}
        totals: {$ref: '#/components/schemas/MenuPeriodStatsResponse'}
        menus:
          type: array
          items: {$ref: '#/components/schemas/UserMenuHistoryEntryResponse'}
    UserMenuHistoryEntryResponse:
      type: object
      properties:
        menuId: {type: integer, format: int64, example: 1}
        startDate: {type: string, format: date}
        endDate: {type: string, format: date}
        stats: {$ref: '#/components/schemas/MenuPeriodStatsResponse'}
...
    MenuPeriodStatsResponse:
      type: object
      properties:
        maxDay:
          $ref: '#/components/schemas/MenuPeriodStatsDayResponse'
        minDay:
          $ref: '#/components/schemas/MenuPeriodStatsDayResponse'
        averageCalories:
          type: number
          description: Average calories per day in the period
          example: 1450.25
        averageCarbohydrates:
          type: number
          description: Average carbohydrates per day in the period
          example: 120.5
        averageProteins:
          type: number
          description: Average proteins per day in the period
          example: 80.1
        averageFats:
          type: number
          description: Average fats per day in the period
          example: 45.75
        moneySpent:
          type: number
          description: Money spent in the period
          example: 32.4
    MenuPeriodStatsDayResponse:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Day date
          example: 2026-06-21
        calories:
          type: number
          description: Calories for that day
          example: 1450.25
...
    NutritionalValuesResponse:
      type: object
      properties:
        calories:
          type: number
          description: Calories
          example: 52
        carbohydrates:
          type: number
          description: Carbohydrates
          example: 14
        proteins:
          type: number
          description: Proteins
          example: 0.3
        fats:
          type: number
          description: Fats
          example: 0.2